requirements.txt
pylint==3.3.1
python-dotenv==1.0.1
Flask==3.0.3
Flask-Cors==5.0.0
SQLAlchemy==2.0.35
pymysql==1.1.1
PyJWT==2.9.0
ldap3==2.9.1
